Update on float32 for matrix_mul.

+++ bigtsne/_matrix_mul/matrix_mul.pxd
@@ -6,13 +6,13 @@
 # cython: language_level=3
 
 cdef void matrix_multiply_fft_1d(
-    double[::1] kernel_tilde,
-    double[:, ::1] w_coefficients,
-    double[:, ::1] out,
+    float[::1] kernel_tilde,
+    float[:, ::1] w_coefficients,
+    float[:, ::1] out,
 )
 
 cdef void matrix_multiply_fft_2d(
-    double[:, ::1] kernel_tilde,
-    double[:, ::1] w_coefficients,
-    double[:, ::1] out,
+    float[:, ::1] kernel_tilde,
+    float[:, ::1] w_coefficients,
+    float[:, ::1] out,
 )
